storage:  Misc corrections and cleanups
 * Make the partial sector logic a little clearer (no functional change)
 * Corrections for debugging messages
 * Also use MAX_VIRT_SECTOR_SIZE in BOUNCE_BUFFER calculations

+++ firmware/common/file.c
@@ -120,7 +120,7 @@ static int flush_cache(struct filestr_desc *file)
     int rc;
     struct filestr_cache *cachep = file->stream.cachep;
 
-    DEBUGF("Flushing dirty sector cache (%lu)\n", cachep->sector);
+    DEBUGF("Flushing dirty sector cache (%llu)\n", (uint64_t)cachep->sector);
 
     if (fat_query_sectornum(&file->stream.fatstr) != cachep->sector)
     {
